Move function, value, environment into transpiler
