+++ src/lib.rs
@@ -7,12 +7,9 @@ extern crate pest;
 extern crate pest_derive;
 
 pub mod ast;
-pub mod environment;
 pub mod error;
-pub mod function;
 pub mod optimizer;
 pub mod parser;
 pub mod transpiler;
-pub mod value;
 
 pub use ast::{YolkProgram, YololProgram};
